Remove all the sample code from SDK. Changed the context file to use the cron expression from global properties.
