--- src/main/amp/config/alfresco/extension/templates/webscripts/helloworld.get.desc.xml
@@ -1,7 +1,0 @@
-<webscript>
-    <shortname>Hello World Sample Webscript</shortname>
-    <description>Hands back a greeting</description>
-    <url>/sample/helloworld</url>
-    <authentication>user</authentication>
-    <format default="html"></format>
-</webscript>
