--- src/main/amp/config/alfresco/extension/templates/webscripts/helloworld.get.html.ftl
@@ -1,1 +1,0 @@
-Message: ${fromJS} ${fromJava}
