--- src/main/amp/config/alfresco/extension/templates/webscripts/helloworld.get.js
@@ -1,1 +1,0 @@
-model["fromJS"] = "Hello World from JS!";
