--- src/main/amp/config/alfresco/module/alf-trash-cleaner/context/service-context.xml
@@ -1,61 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<!DOCTYPE beans PUBLIC '-//SPRING//DTD BEAN//EN' 'http://www.springframework.org/dtd/spring-beans.dtd'>
-<!--
-    Licensed to the Apache Software Foundation (ASF) under one or more
-    contributor license agreements.  See the NOTICE file distributed with
-    this work for additional information regarding copyright ownership.
-    The ASF licenses this file to You under the Apache License, Version 2.0
-    (the "License"); you may not use this file except in compliance with
-    the License.  You may obtain a copy of the License at
-    	
-    http://www.apache.org/licenses/LICENSE-2.0
-    	
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<beans>
-
-    <!-- A simple class that is initialized by Spring -->
-    <bean id="org.ootb.exampleBean" class="org.ootb.demoamp.Demo" init-method="init" />
-
-    <!-- A simple module component that will be executed once -->
-    <bean id="org.ootb.exampleComponent" class="org.ootb.demoamp.DemoComponent" parent="module.baseComponent" >
-        <property name="moduleId" value="${project.artifactId}" />  <!-- See module.properties -->
-        <property name="name" value="exampleComponent" />
-        <property name="description" value="A demonstration component" />
-        <property name="sinceVersion" value="2.0" />
-        <property name="appliesFromVersion" value="2.0" />
-        <property name="nodeService" ref="NodeService" />
-        <property name="nodeLocatorService" ref="nodeLocatorService" />
-    </bean>
-    
-
-  <bean id="trashcanCleaner" class="org.alfresco.util.CronTriggerBean">
-        <property name="jobDetail">
-            <bean id="tempFileCleanerJobDetail" class="org.springframework.scheduling.quartz.JobDetailBean">
-                <property name="jobClass">
-                    <value>org.ootb.trashcleaner.TrashcanCleanerJob</value>
-                </property>
-                <property name="jobDataAsMap">
-                        <map>
-                                <entry key="nodeService" value-ref="nodeService" />
-                                <entry key="transactionService" value-ref="transactionService" />
-                                <entry key="authenticationComponent" value-ref="authenticationComponent" />
-                                <entry key="jobLockService" value-ref="jobLockService" />
-                                <entry key="trashcan.daysToKeep" value="${trashcan.daysToKeep}" />
-                                <entry key="trashcan.deleteBatchCount" value="${trashcan.deleteBatchCount}" />
-                        </map>
-                </property>
-            </bean>
-        </property>
-        <property name="scheduler">
-            <ref bean="schedulerFactory" />
-        </property>
-        <property name="cronExpression">
-                <value>${trashcan.cleaner.cron}</value>
-        </property>
-    </bean>
-</beans>
