--- src/main/amp/config/alfresco/module/alf-trash-cleaner/module-context.xml
@@ -1,26 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<!--
-	Licensed to the Apache Software Foundation (ASF) under one or more
-	contributor license agreements. 
-	The ASF licenses this file to You under the Apache License, Version 2.0
-	(the "License"); you may not use this file except in compliance with
-	the License.  You may obtain a copy of the License at
-	
-	http://www.apache.org/licenses/LICENSE-2.0
-	
-	Unless required by applicable law or agreed to in writing, software
-	distributed under the License is distributed on an "AS IS" BASIS,
-	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-	See the License for the specific language governing permissions and
-	limitations under the License.
-	
--->
-<!DOCTYPE beans PUBLIC '-//SPRING//DTD BEAN//EN' 'http://www.springframework.org/dtd/spring-beans.dtd'>
-
-<beans>
-	<!-- This is filtered by Maven at build time, so that module name is single sourced. -->
-    <import resource="classpath:alfresco/module/${project.artifactId}/context/service-context.xml" />
-    <import resource="classpath:alfresco/module/${project.artifactId}/context/bootstrap-context.xml" />
-	<import resource="classpath:alfresco/module/${project.artifactId}/context/webscript-context.xml" />
-
-</beans>
