+++ src/main/amp/config/alfresco/module/alf-trashcan-cleaner/alfresco-global.properties
@@ -3,9 +3,6 @@
 ## System Administrators can override these values in environment specific configurations in
 ## alfresco/tomcat/shared/classes/alfresco-global.properties.
 ## 
-
-##trashcan.cleaner.cron=0 0 4 * * ?
-trashcan.cleaner.cron=0 * 22 * * ?
-trashcan.cleaner.protected.day=1
-trashcan.daysToKeep=0
+trashcan.daysToKeep=-1
 trashcan.deleteBatchCount=1000
+trashcan.cron=0 0/2 * * * ?
