+++ src/main/amp/config/alfresco/module/alf-trashcan-cleaner/context/service-context.xml
@@ -1,0 +1,47 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<!DOCTYPE beans PUBLIC '-//SPRING//DTD BEAN//EN' 'http://www.springframework.org/dtd/spring-beans.dtd'>
+<!--
+    Licensed to the Apache Software Foundation (ASF) under one or more
+    contributor license agreements.  See the NOTICE file distributed with
+    this work for additional information regarding copyright ownership.
+    The ASF licenses this file to You under the Apache License, Version 2.0
+    (the "License"); you may not use this file except in compliance with
+    the License.  You may obtain a copy of the License at
+    	
+    http://www.apache.org/licenses/LICENSE-2.0
+    	
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+-->
+<beans>
+
+    <bean id="tempFileCleanerJobDetail" class="org.springframework.scheduling.quartz.JobDetailBean">
+            <property name="jobClass">
+                <value>com.ootb.trashcan.TrashcanCleanerJob</value>
+            </property>
+            <property name="jobDataAsMap">
+			<map>
+				<entry key="nodeService" value-ref="nodeService" />
+				<entry key="transactionService" value-ref="transactionService" />
+				<entry key="authenticationComponent" value-ref="authenticationComponent" />
+				<entry key="jobLockService" value-ref="jobLockService" />
+				<entry key="daysToKeep" value="${trashcan.daysToKeep}" />
+				<entry key="deleteBatchCount" value="${trashcan.deleteBatchCount}" />
+			</map>
+			</property>
+    </bean>
+    <bean id="trashcanCleaner" class="org.alfresco.util.CronTriggerBean">
+        <property name="jobDetail">
+			<ref bean="tempFileCleanerJobDetail" />
+        </property>
+        <property name="scheduler">
+            <ref bean="schedulerFactory" />
+        </property>
+        <property name="cronExpression">
+			<value>${trashcan.cron}</value>
+        </property>
+    </bean>
+</beans>
