+++ src/main/amp/config/alfresco/module/alf-trashcan-cleaner/log4j.properties
@@ -38,5 +38,4 @@
 #    
 #-----------------------------------------------------------------------
 log4j.logger.org.alfresco.demoamp.DemoComponent=${module.log.level}
-log4j.logger.org.ootb.trashcleaner.TrashcanCleaner=info
-log4j.logger.org.ootb.trashcleaner=DEBUG
+log4j.logger.com.ootb.trashcan.TrashcanCleaner=DEBUG
