+++ src/main/amp/config/alfresco/module/alf-trashcan-cleaner/model/content-model.xml
