+++ src/main/amp/config/alfresco/module/alf-trashcan-cleaner/model/workflow-model.xml
