--- src/main/amp/web/licenses/README-licenses.txt
@@ -1,2 +1,0 @@
-This folder (root in the AMP) gets mapped automagically in WEB-INF/licenses
-by the MMT or the alfresco-maven-plugin
