--- src/main/java/org/ootb/demoamp/Demo.java
@@ -1,15 +1,0 @@
-package org.ootb.demoamp;
-
-/**
- * This class does nothing except dump some output to <i>system.out</i>.
- * This is a sample taken from Maven Alfresco SDK
- * 
- * @author Derek Hulley
- */
-public class Demo
-{
-	public void init()
-	{
-		System.out.println("SDK Demo AMP class has been loaded");
-	}
-}
